requirements.txt
SQLAlchemy==2.0.30
psycopg2-binary>=2.9.0
fastapi==0.115.12
passlib==1.7.4
pydantic==2.11.4
pydantic_settings==2.9.1
python_jose==3.3.0
python-multipart